fix: not-found 페이지 헤더 추가
- sentry 로깅 NotFoundLogger 컴포넌트로 분리
